# openapi.json
{
  "components": {
    "schemas": {
...
      "AssetPrepare": {
        "properties": {
          "allow_cloud": {
            "default": false,
            "title": "Allow Cloud",
            "type": "boolean"
          },
          "file": {
            "default": "",
            "maxLength": 500,
            "title": "File",
            "type": "string"
          },
          "max_credits": {
            "default": 0,
            "maximum": 120.0,
            "minimum": 0.0,
            "title": "Max Credits",
            "type": "integer"
          },
          "max_triangles": {
            "default": 30000,
            "maximum": 100000.0,
            "minimum": 100.0,
            "title": "Max Triangles",
            "type": "integer"
          },
          "source": {
            "default": "local",
            "enum": [
              "local",
              "meshy-task",
              "meshy-generate"
            ],
            "title": "Source",
            "type": "string"
          },
          "task_id": {
            "default": "",
            "maxLength": 64,
            "pattern": "^[a-zA-Z0-9-]*$",
            "title": "Task Id",
            "type": "string"
          },
          "task_type": {
            "default": "image-to-3d",
            "enum": [
              "image-to-3d",
              "text-to-3d"
            ],
            "title": "Task Type",
            "type": "string"
          }
        },
        "title": "AssetPrepare",
        "type": "object"
      },
      "AssetUpload": {
        "properties": {
          "data": {
            "maxLength": 73400320,
            "minLength": 1,
            "title": "Data",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "AssetUpload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UnrealImport": {
        "properties": {
          "file": {
            "maxLength": 500,
            "minLength": 1,
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "UnrealImport",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local AI production API for Beast Studio (127.0.0.1:8787, this machine only). Contract generated from the live FastAPI app \u2014 see AGENT_ACCESS.md for operational rules (credit/privacy defaults, backend-sharing etiquette) that this schema does not itself encode. Sync endpoints return results directly; async endpoints return {id} and are polled via GET /api/run/{run_id} or streamed via GET /api/events/{run_id} (SSE).",
    "title": "Beast Studio API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/assets/prepare": {
      "post": {
        "operationId": "prepare_asset_api_assets_prepare_post",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetPrepare"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prepare Asset"
      }
    },
...
    "/api/assets/unreal": {
      "post": {
        "operationId": "unreal_import_api_assets_unreal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnrealImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unreal Import"
      }
    },
    "/api/assets/upload": {
      "post": {
        "operationId": "upload_asset_api_assets_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Asset"
      }
    },
...
  },
  "servers": [
    {
      "description": "local Beast Studio server",
      "url": "http://127.0.0.1:8787"
    }
  ]
}